feat: refactor MessageQualityBarChart and integrate message quality fetching in Analysis page
